Delete AddTaskCommand.java
