+++ docs/team/weenleen.md
@@ -1,0 +1,44 @@
+---
+layout: page
+title: Win Lynn's Project Portfolio Page
+---
+
+### Project: tApp
+
+Given below are my contributions to the project.
+
+* **Code contributed:** [RepoSense link](https://nus-cs2103-ay2122s1.github.io/tp-dashboard/?search=weenleen)
+
+* **New Feature**: Added the Task class
+    * What it does: allows the user to undo all previous commands one at a time. Preceding undo commands can be reversed by using the redo command.
+    * Justification: This feature improves the product significantly because a user can make mistakes in commands and the app should provide a convenient way to rectify them.
+    * Highlights: This enhancement affects existing commands and commands to be added in future. It required an in-depth analysis of design alternatives. The implementation too was challenging as it required changes to existing commands.
+    * Credits: *{mention here if you reused any code/ideas from elsewhere or if a third-party library is heavily used in the feature so that a reader can make a more accurate judgement of how much effort went into the feature}*
+
+* **New Feature**: Added a history command that allows the user to navigate to previous commands using up/down keys.
+
+* **Project management**:
+    * Managed releases `v1.3` - `v1.5rc` (3 releases) on GitHub
+
+* **Enhancements to existing features**:
+    * Updated the GUI color scheme (Pull requests [\#33](), [\#34]())
+    * Wrote additional tests for existing features to increase coverage from 88% to 92% (Pull requests [\#36](), [\#38]())
+
+* **Documentation**:
+    * User Guide:
+        * Added documentation for the features `delete` and `find` [\#72]()
+        * Did cosmetic tweaks to existing documentation of features `clear`, `exit`: [\#74]()
+    * Developer Guide:
+        * Added implementation details of the `delete` feature.
+
+* **Community**:
+    * PRs reviewed (with non-trivial review comments): [\#12](), [\#32](), [\#19](), [\#42]()
+    * Contributed to forum discussions (examples: [1](), [2](), [3](), [4]())
+    * Reported bugs and suggestions for other teams in the class (examples: [1](), [2](), [3]())
+    * Some parts of the history feature I added was adopted by several other class mates ([1](), [2]())
+
+* **Tools**:
+    * Integrated a third party library (Natty) to the project ([\#42]())
+    * Integrated a new Github plugin (CircleCI) to the team repo
+
+* _{you can add/remove categories in the list above}_
