+++ src/test/java/seedu/address/logic/commands/AddGroupCommandTest.java
@@ -1,28 +1,16 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.group.Group;
-import seedu.address.model.student.Student;
-import seedu.address.model.task.Task;
 import seedu.address.testutil.GroupBuilder;
 
 public class AddGroupCommandTest {
@@ -34,7 +22,7 @@ public class AddGroupCommandTest {
 
     @Test
     public void execute_groupAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingGroupAdded modelStub = new ModelStubAcceptingGroupAdded();
+        CommandTestUtil.ModelStubAcceptingGroupAdded modelStub = new CommandTestUtil.ModelStubAcceptingGroupAdded();
         Group validGroup = new GroupBuilder().build();
 
         CommandResult commandResult = new AddGroupCommand(validGroup).execute(modelStub);
@@ -47,7 +35,7 @@ public class AddGroupCommandTest {
     public void execute_duplicateGroup_throwsCommandException() {
         Group validGroup = new GroupBuilder().build();
         AddGroupCommand addGroupCommand = new AddGroupCommand(validGroup);
-        AddGroupCommandTest.ModelStub modelStub = new ModelStubWithGroup(validGroup);
+        CommandTestUtil.ModelStub modelStub = new CommandTestUtil.ModelStubWithGroup(validGroup);
 
         assertThrows(CommandException.class,
                 AddGroupCommand.MESSAGE_DUPLICATE_GROUP, () -> addGroupCommand.execute(modelStub));
@@ -75,233 +63,6 @@ public class AddGroupCommandTest {
 
         // different task -> returns false
         assertFalse(addG1Command.equals(addG2Command));
-    }
-
-    /**
-     * A default model stub that has all methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addStudent(Student student) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void clearTasks() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasStudent(Student student) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasAnotherStudent(Student student, Student toIgnore) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteStudent(Student target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setStudent(Student target, Student editedStudent) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getStudentAttendance(Student target, int week) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void markStudentAttendance(Student target, int week) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getStudentParticipation(Student target, int week) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void markStudentParticipation(Student target, int week) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Student> getFilteredStudentList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredStudentList(Predicate<Student> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasTask(Task task) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteTask(Task target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void toggleTaskIsDone(Task target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addTask(Task task) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setTask(Task target, Task editedTask) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Task> getFilteredTaskList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredTaskList(Predicate<Task> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public DisplayType getDisplayType() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasGroup(Group group) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteGroup(Group target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteMember(Student student, Group group) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addGroup(Group group) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGroup(Group target, Group editedGroup) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Group> getFilteredGroupList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredGroupList(Predicate<Group> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addMember(Student target, Group group) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
-     * A Model stub that contains a single task.
-     */
-    private class ModelStubWithGroup extends AddGroupCommandTest.ModelStub {
-        private final Group group;
-
-        ModelStubWithGroup(Group group) {
-            requireNonNull(group);
-            this.group = group;
-        }
-
-        @Override
-        public boolean hasGroup(Group group) {
-            requireNonNull(group);
-            return this.group.isSameGroup(group);
-        }
-    }
-
-    /**
-     * A Model stub that always accept the group being added.
-     */
-    private class ModelStubAcceptingGroupAdded extends AddGroupCommandTest.ModelStub {
-        final ArrayList<Group> groupsAdded = new ArrayList<>();
-
-        @Override
-        public boolean hasGroup(Group group) {
-            requireNonNull(group);
-            return groupsAdded.stream().anyMatch(group::isSameGroup);
-        }
-
-        @Override
-        public void addGroup(Group group) {
-            requireNonNull(group);
-            groupsAdded.add(group);
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            return new AddressBook();
-        }
     }
 
 }
