+++ src/test/java/seedu/address/logic/commands/DeleteStudentCommandTest.java
@@ -24,7 +24,7 @@ import seedu.address.model.student.Student;
  */
 public class DeleteStudentCommandTest {
 
-    private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+    private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
 
     @Test
     public void execute_validIndexUnfilteredList_success() {
