+++ src/test/java/seedu/address/logic/commands/MarkStudentAttCommandTest.java
@@ -32,7 +32,7 @@ public class MarkStudentAttCommandTest {
     private static final Integer[] MARK_ATTENDANCE_ABSENT_EXPECTED = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
     private static final int FIRST_WEEK = Attendance.FIRST_WEEK_OF_SEM;
 
-    private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+    private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
 
     @Test
     public void execute_markStudentPresentUnfilteredList_success() {
