+++ src/test/java/seedu/address/logic/parser/DeleteStudentCommandParserTest.java
@@ -18,7 +18,7 @@ import seedu.address.logic.commands.DeleteStudentCommand;
  */
 public class DeleteStudentCommandParserTest {
 
-    private DeleteStudentCommandParser parser = new DeleteStudentCommandParser();
+    private final DeleteStudentCommandParser parser = new DeleteStudentCommandParser();
 
     @Test
     public void parse_validArgs_returnsDeleteCommand() {
