+++ src/test/java/seedu/address/logic/parser/EditTaskCommandParserTest.java
@@ -36,7 +36,7 @@ public class EditTaskCommandParserTest {
     private static final String MESSAGE_INVALID_FORMAT =
             String.format(MESSAGE_INVALID_COMMAND_FORMAT, EditTaskCommand.MESSAGE_USAGE);
 
-    private EditTaskCommandParser parser = new EditTaskCommandParser();
+    private final EditTaskCommandParser parser = new EditTaskCommandParser();
 
     @Test
     public void parse_missingParts_failure() {
